+++ config/src/main/java/org/hyperledger/besu/config/GenesisConfigOptions.java
@@ -90,6 +90,10 @@ public interface GenesisConfigOptions {
 
   OptionalLong getMergeNetSplitBlockNumber();
 
+  OptionalLong getShanghaiTime();
+
+  OptionalLong getCancunTime();
+
   OptionalLong getShandongBlockNumber();
 
   Optional<Wei> getBaseFeePerGas();
@@ -100,7 +104,9 @@ public interface GenesisConfigOptions {
 
   Optional<Hash> getTerminalBlockHash();
 
-  List<Long> getForks();
+  List<Long> getForkBlockNumbers();
+
+  List<Long> getForkBlockTimestamps();
 
   /**
    * Block number for the Dao Fork, this value is used to tell node to connect with peer that did