+++ consensus/merge/src/main/java/org/hyperledger/besu/consensus/merge/MergeContext.java
@@ -65,4 +65,12 @@ public interface MergeContext extends ConsensusContext {
   void putPayloadById(final PayloadIdentifier payloadId, final Block block);
 
   Optional<Block> retrieveBlockById(final PayloadIdentifier payloadId);
+
+  default void setIsChainPruningEnabled(final boolean isChainPruningEnabled) {}
+
+  default boolean isChainPruningEnabled() {
+    return false;
+  }
+
+  boolean isCheckpointPostMergeSync();
 }