+++ ethereum/evmtool/src/main/java/org/hyperledger/besu/evmtool/EvmTool.java
@@ -15,11 +15,14 @@
  */
 package org.hyperledger.besu.evmtool;
 
+import org.hyperledger.besu.crypto.SignatureAlgorithmFactory;
+
 import picocli.CommandLine;
 
 public final class EvmTool {
 
   public static void main(final String... args) {
+    SignatureAlgorithmFactory.setDefaultInstance();
 
     final EvmToolCommand evmToolCommand = new EvmToolCommand();