+++ evm/src/main/java/org/hyperledger/besu/evm/frame/MessageFrame.java
@@ -21,9 +21,11 @@ import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.evm.Code;
+import org.hyperledger.besu.evm.code.CodeSection;
 import org.hyperledger.besu.evm.internal.FixedStack.UnderflowException;
 import org.hyperledger.besu.evm.internal.MemoryEntry;
 import org.hyperledger.besu.evm.internal.OperandStack;
+import org.hyperledger.besu.evm.internal.ReturnStack;
 import org.hyperledger.besu.evm.internal.StorageEntry;
 import org.hyperledger.besu.evm.log.Log;
 import org.hyperledger.besu.evm.operation.Operation;
@@ -201,8 +203,10 @@ public class MessageFrame {
   private final Function<Long, Hash> blockHashLookup;
   private final int maxStackSize;
   private int pc;
+  private int section;
   private final Memory memory;
   private final OperandStack stack;
+  private final ReturnStack returnStack;
   private Bytes output;
   private Bytes returnData;
   private final boolean isStatic;
@@ -278,8 +282,10 @@ public class MessageFrame {
     this.blockHashLookup = blockHashLookup;
     this.maxStackSize = maxStackSize;
     this.pc = 0;
+    this.section = 0;
     this.memory = new Memory();
     this.stack = new OperandStack(maxStackSize);
+    this.returnStack = new ReturnStack();
     this.output = Bytes.EMPTY;
     this.returnData = Bytes.EMPTY;
     this.logs = new ArrayList<>();
@@ -339,6 +345,52 @@ public class MessageFrame {
    */
   public void setPC(final int pc) {
     this.pc = pc;
+  }
+
+  /**
+   * Return the current code section. Always zero for legacy code.
+   *
+   * @return the current code section
+   */
+  public int getSection() {
+    return section;
+  }
+
+  public ExceptionalHaltReason callFunction(final int calledSection) {
+    CodeSection info = code.getCodeSection(calledSection);
+    if (info == null) {
+      return ExceptionalHaltReason.CODE_SECTION_MISSING;
+    } else if (stack.size() + info.getMaxStackHeight() > maxStackSize) {
+      return ExceptionalHaltReason.TOO_MANY_STACK_ITEMS;
+    } else {
+      returnStack.push(
+          new ReturnStack.ReturnStackItem(section, pc + 2, stack.size() - info.getInputs()));
+      pc = -1; // will be +1ed at end of operations loop
+      this.section = calledSection;
+      return null;
+    }
+  }
+
+  public ExceptionalHaltReason jumpFunction(final int section) {
+    CodeSection thisInfo = code.getCodeSection(this.section);
+    CodeSection info = code.getCodeSection(section);
+    if (info == null) {
+      return ExceptionalHaltReason.CODE_SECTION_MISSING;
+    } else if (thisInfo.getOutputs() != info.getOutputs()) {
+      return ExceptionalHaltReason.MISMATCHED_CODE_SECTION_OUTPUTS;
+    } else {
+      pc = -1; // will be +1ed at end of operations loop
+      this.section = section;
+      return null;
+    }
+  }
+
+  public void returnFunction() {
+    CodeSection thisInfo = code.getCodeSection(this.section);
+    var returnInfo = returnStack.pop();
+    stack.preserveTop(returnInfo.getStackHeight(), thisInfo.getOutputs());
+    this.pc = returnInfo.getPC();
+    this.section = returnInfo.getCodeSectionIndex();
   }
 
   /** Deducts the remaining gas. */
@@ -583,11 +635,11 @@ public class MessageFrame {
    */
   public MutableBytes readMutableMemory(
       final long offset, final long length, final boolean explicitMemoryRead) {
-    final MutableBytes value = memory.getMutableBytes(offset, length);
+    final MutableBytes memBytes = memory.getMutableBytes(offset, length);
     if (explicitMemoryRead) {
-      setUpdatedMemory(offset, value);
+      setUpdatedMemory(offset, memBytes);
     }
-    return value;
+    return memBytes;
   }
 
   /**