+++ evm/src/main/java/org/hyperledger/besu/evm/operation/InvalidOperation.java
@@ -31,7 +31,7 @@ public class InvalidOperation extends AbstractOperation {
   }
 
   public InvalidOperation(final int opcode, final GasCalculator gasCalculator) {
-    super(opcode, "INVALID", -1, -1, 1, gasCalculator);
+    super(opcode, "INVALID", -1, -1, gasCalculator);
     invalidResult = new OperationResult(0L, ExceptionalHaltReason.INVALID_OPERATION);
   }