+++ evm/src/main/java/org/hyperledger/besu/evm/operation/JumpOperation.java
@@ -28,7 +28,7 @@ public class JumpOperation extends AbstractFixedCostOperation {
   private final OperationResult jumpResponse;
 
   public JumpOperation(final GasCalculator gasCalculator) {
-    super(0x56, "JUMP", 2, 0, 1, gasCalculator, gasCalculator.getMidTierGasCost());
+    super(0x56, "JUMP", 2, 0, gasCalculator, gasCalculator.getMidTierGasCost());
     invalidJumpResponse =
         new Operation.OperationResult(gasCost, ExceptionalHaltReason.INVALID_JUMP_DESTINATION);
     jumpResponse = new OperationResult(gasCost, null, 0);