+++ evm/src/main/java/org/hyperledger/besu/evm/operation/XorOperation.java
@@ -25,7 +25,7 @@ public class XorOperation extends AbstractFixedCostOperation {
   static final OperationResult xorSuccess = new OperationResult(3, null);
 
   public XorOperation(final GasCalculator gasCalculator) {
-    super(0x18, "XOR", 2, 1, 1, gasCalculator, gasCalculator.getVeryLowTierGasCost());
+    super(0x18, "XOR", 2, 1, gasCalculator, gasCalculator.getVeryLowTierGasCost());
   }
 
   @Override